+++ src/components/AmiibosParams.tsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 import { useAppSelector } from '../redux/hooks';
-import { Amiibo } from '../interfaces/amiiboInterface';
+import { Amiibo } from '../types/Amiibo';
 import { Link } from 'react-router-dom';
 
 import styled from '@emotion/styled';
