+++ src/features/amiibo/getAllAmiibo.ts
