+++ src/redux/store.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { allAmiiboSliceReducer } from './getAllAmiibo';
-import { userSliceReducer } from './userSlice';
+import { allAmiiboSliceReducer } from '../features/amiibo/getAllAmiibo';
+import { userSliceReducer } from '../features/user/userSlice';
 
 //
 export const store = configureStore({
