+++ src/types/Amiibo.ts
